+++ core/engine.js
@@ -1,4 +1,4 @@
-import { $primarySymbols, $revPrimarySymbols, $generalPunctuation, $revGeneralPunctuation, $currencySymbols, $revCurrencySymbols, $latinSymbols, $revLatinSymbols } from "../char/characters.js";
+import { $primarySymbols, $revPrimarySymbols, $generalPunctuation, $revGeneralPunctuation, $currencySymbols, $revCurrencySymbols, $latinSymbols, $revLatinSymbols } from "../cipher/characters.js";
 
 const encryptionAlgorithm = () => `abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890 ${$primarySymbols}${$generalPunctuation}${$currencySymbols}${$latinSymbols}`;
 const decryptionAlgorithm = () => `qwertyuiopasdfghjklzxcvbnmQWERTYUIOPASDFGHJKLZXCVBNM0987654321 ${$revLatinSymbols}${$revCurrencySymbols}${$revGeneralPunctuation}${$revPrimarySymbols}`;
